+++ examples/multi-tenancy-appwrite/src/components/select/index.tsx
@@ -3,6 +3,7 @@ import { useGetToPath, useGo, useParsed } from "@refinedev/core";
 import { Select } from "antd";
 
 import { IStore } from "../../interfaces";
+import { resources } from "../../utility/appwriteClient";
 
 export const StoreSelect: React.FC = () => {
     const getToPath = useGetToPath();
@@ -10,7 +11,7 @@ export const StoreSelect: React.FC = () => {
     const { resource, action, params } = useParsed<{ tenant: string }>();
 
     const { selectProps: storeSelectProps } = useSelect<IStore>({
-        resource: "stores",
+        resource: resources.stores,
         optionLabel: "title",
         optionValue: "id",
         meta: {